Fix for numerical error
Fix for an issue where if an envelope was re-triggered when it was at the very end of the attack stage it could get locked and never complete. This was because c_inc_ could become so small (due to the delta between the current value and the target value being very small, and then divided by the sample rate) that when added to val the value of val does not change (due to floating point inaccuracies). Ensure c_inc_ is at least as big as std epsilon, so that it will always cause val to change.